render all artist page
