+++ views/all-artwork.handlebars
