+++ views/homepage.handlebars
@@ -1,7 +1,7 @@
         <section class="p-3 mb-3 bg-dark flex-row align-center">
             <h2 class="text-light col-12 col-lg-5">Add To Our Catalog.</h2>
             <div class="col-12 col-lg-7 flex-row text-center justify-space-around align-center">
-                <a class="btn col-12 col-md-auto mx-2" href="/">View Artwork</a>
+                <a class="btn col-12 col-md-auto mx-2" href="/all-artwork">View Artwork</a>
                 <!-- do we nee another view?<a class="btn col-12 col-md-auto mx-2" href="/zookeepers">View </a>--->
             </div>
         </section>
